requirements.txt
astroid==3.1.0
Babel==2.15.0
backports-datetime-fromisoformat==2.0.1
beautifulsoup4==4.12.3
black==24.4.2
blessed==1.20.0
bpython==0.24
Brotli==1.1.0
bs4==0.0.2
cchardet==2.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colored==2.2.4
courlan==1.1.0
cssselect==1.2.0
curtsies==0.4.2
cwcwidth==0.1.9
dateparser==1.2.0
dill==0.3.8
EbookLib==0.18
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
Gooey==1.0.8.1
greenlet==3.0.3
htmldate==1.8.1
idna==3.7
isort==5.13.2
jieba3k==0.35.1
joblib==1.4.2
jusText==3.0.1
lxml==5.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
psutil==5.9.8
py3langid==0.2.2
pycurl==7.45.3
Pygments==2.18.0
pygtrie==2.5.0
pylint==3.1.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
requests-file==2.0.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tinysegmenter==0.3
tld==0.13
tldextract==5.1.2
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
wxPython==4.2.1
